fix: add tag_override for script in flows
