+++ frontend/src/lib/components/TestJobLoader.svelte
@@ -91,6 +91,8 @@
 
 	export async function getLogs() {
 		if (job) {
+			console.log('getLogs')
+
 			const getUpdate = await JobService.getJobUpdates({
 				workspace: workspace!,
 				id: job.id,
@@ -172,6 +174,8 @@
 			try {
 				if (job && `running` in job) {
 					const offset = logOffset == 0 ? (job.logs?.length ? job.logs?.length + 1 : 0) : logOffset
+					console.log('getLogs')
+
 					let previewJobUpdates = await JobService.getJobUpdates({
 						workspace: workspace!,
 						id,
