+++ frontend/src/lib/components/common/confirmationModal/ConfirmationModal.svelte
@@ -54,7 +54,7 @@
 					class={classNames(
 						'relative transform overflow-hidden rounded-lg bg-surface px-4 pt-5 pb-4 text-left shadow-xl transition-all sm:my-8 sm:w-full sm:max-w-lg sm:p-6',
 						open
-							? 'ease-out duration-300 opacity-80 translate-y-0 sm:scale-100'
+							? 'ease-out duration-300 opacity-100 translate-y-0 sm:scale-100'
 							: 'ease-in duration-200 opacity-0 translate-y-4 sm:translate-y-0 sm:scale-95'
 					)}
 				>
